# Use `Len64` from "math/bits" in `Deg`